complete shared schedule, make other housekeeping changes
