+++ Dental_Surgery/Pages/Admin2/Appointments/Index.cshtml
@@ -4,10 +4,11 @@
 @{
     ViewData["Title"] = "Appointments";
 }
-<h2>Appointments</h2>
-<div class="container mt-5">
+
+<div class="container mt-4">
+    <h2>Appointments</h2>
     <div class="d-flex justify-content-between align-items-center mb-4">  
-        <a asp-page="/Receptionist/Index" class="btn btn-primary">Create New Appointment</a>
+        <a asp-page="./Create" class="btn btn-primary">Create New Appointment</a>
     </div>
 
     <table class="table table-striped table-bordered table-white">
