+++ Dental_Surgery/Pages/Admin2/Patients/Index.cshtml
@@ -5,9 +5,9 @@
     ViewData["Title"] = "Patients";
 }
 
-<div class="container mt-5">
+<div class="container mt-4">
+    <h2>Patients</h2>
     <div class="d-flex justify-content-between align-items-center mb-4">
-        <h2>Patients</h2>
         <a asp-page="Create" class="btn btn-primary">Add New Patient</a>
     </div>
