+++ Dental_Surgery/Pages/Bills/Index.cshtml
@@ -4,8 +4,8 @@
     ViewData["Title"] = "Bill Generation";
 }
 
-<div class="container mt-5">
-    <h2 class="mb-4">@ViewData["Title"]</h2>
+<div class="container mt-4">
+    <h2>Bill Generation</h2>
 
     <div class="card shadow-sm">
         <div class="card-body">
