+++ Dental_Surgery/Pages/Dentists/Index.cshtml
@@ -1,60 +1,5 @@
 ﻿@page
 @model Dental_Surgery.Pages.Dentists.IndexModel
 @{
-    ViewData["Title"] = "Daily Schedule";
-    var hour = DateTime.Now.Hour;
-    var greeting = hour < 12 ? "Good morning" : (hour < 16 ? "Good afternoon" : "Good evening");
-}
 
-<h2>@greeting, @Model.CurrentDentist.FirstName</h2>
-
-<div style="margin-top:10px; display:flex; justify-content: space-between; align-items:center;">
-    <strong>@Model.ScheduleDate.ToString("dddd, MMM dd yyyy")</strong>
-    
-    <a asp-page="./Index"
-       asp-route-ScheduleDate="@Model.ScheduleDate.AddDays(-1).ToString("yyyy-MM-dd")"
-       class="btn btn-secondary">&larr; Previous Day</a>
-    <a asp-page="./Index"
-       asp-route-ScheduleDate="@Model.ScheduleDate.AddDays(1).ToString("yyyy-MM-dd")"
-       class="btn btn-secondary">Next Day &rarr;</a>
-</div>
-
-@if (Model.CurrentDentist != null)
-{
-    @if (Model.DailyAppointments.Any())
-    {
-        <div class="container mt-4">
-            <div class="table-responsive shadow-sm">
-                <table class="table table-striped table-bordered align-middle table-white">
-                    <thead class="table-dark">
-                        <tr>
-                            <th>Time</th>
-                            <th>Patient</th>
-                            <th>Treatment</th>
-                            <th>Details</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        @foreach (var appt in Model.DailyAppointments)
-                        {
-                            <tr>
-                                <td>@appt.AppointmentDate.ToString("HH:mm")</td>
-                                <td>@appt.Patient?.FirstName @appt.Patient?.LastName</td>
-                                <td>@appt.Treatment?.Name</td>
-                                <td><a asp-page="/Admin2/Appointments/Details" asp-route-id="@appt.AppointmentId" class="btn btn-info btn-sm">Details</a></td>
-                            </tr>
-                        }
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    }
-    else
-    {
-        <p>No appointments for this day.</p>
-    }
-}
-else
-{
-    <p>Unable to load schedule. Dentist not found.</p>
 }
