+++ Dental_Surgery/Pages/Index.cshtml.cs
@@ -12,9 +12,9 @@ namespace Dental_Surgery.Pages
             _logger = logger;
         }
 
-        public void OnGet()
+        public IActionResult OnGet()
         {
-
-        }
+			return RedirectToPage("/Login");
+		}
     }
 }
