+++ Dental_Surgery/Pages/Shared/_Layout.cshtml
@@ -67,20 +67,20 @@
 					@if (roles.Contains("Dentist"))
 					{
 						<li class="navbar__item">
-							<a class="navbar__link" asp-area="" asp-page="/Dentists/Index">
-								<i data-feather="clipboard"></i><span>Home</span>
+							<a class="navbar__link" asp-area="" asp-page="/Shared/Schedule">
+								<i data-feather="home"></i><span>Home</span>
 							</a>
 						</li>
 					}
 					@if (roles.Contains("Receptionist"))
 					{
 						<li class="navbar__item">
-							<a class="navbar__link" asp-area="" asp-page="/Index">
+							<a class="navbar__link" asp-area="" asp-page="/Shared/Schedule">
 								<i data-feather="home"></i><span>Home</span>
 							</a>
 						</li>
 						<li class="navbar__item">
-							<a class="navbar__link" asp-area="" asp-page="/Receptionist/Index">
+							<a class="navbar__link" asp-area="" asp-page="/Admin2/Appointments/Create">
 								<i data-feather="book-open"></i><span>Create Appointment</span>
 							</a>
 						</li>
@@ -91,10 +91,10 @@
 						</a>
 					</li>
 					<li class="navbar__item">
-    <a class="navbar__link" asp-area="" asp-page="/Bills/Index">
-        <i data-feather="dollar-sign"></i><span>Bills</span>
-    </a>
-</li>
+						<a class="navbar__link" asp-area="" asp-page="/Bills/Index">
+							<i data-feather="dollar-sign"></i><span>Bills</span>
+						</a>
+					</li>
 					<li class="navbar__item">
 						<a class="navbar__link" asp-area="" asp-page="/Admin2/Patients/Index">
 							<i data-feather="users"></i><span>Patients</span>
